Begin server conversion from Python to C++
The first step towards a C++ server is a simple implementation of the belt
buckle command class and a few methods to produce a serial string. I've also
started on the main server loop. It does nothing currently except use an
imprecise time/sleep setup to keep it ticking at ~60 UPS.